# openapi.json
{"openapi":"3.0.0","paths":{"/api/v2204/trustAnchor":{"post":{"operationId":"TrustAnchorController_findTrustAnchor","summary":"Search for a TrustAnchor certificate in the registry","parameters":[],"requestBody":{"required":true,"description":"Search for a TrustAnchor certificate in the registry","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrustAnchorRequestDto"},"examples":{"trustAnchor":{"summary":"Example Certificate","value":{"certificate":"MIIEvTCCA6WgAwIBAgIBADANBgkqhkiG9w0BAQUFADB/MQswCQYDVQQGEwJFVTEnMCUGA1UEChMeQUMgQ2FtZXJmaXJtYSBTQSBDSUYgQTgyNzQzMjg3MSMwIQYDVQQLExpodHRwOi8vd3d3LmNoYW1iZXJzaWduLm9yZzEiMCAGA1UEAxMZQ2hhbWJlcnMgb2YgQ29tbWVyY2UgUm9vdDAeFw0wMzA5MzAxNjEzNDNaFw0zNzA5MzAxNjEzNDRaMH8xCzAJBgNVBAYTAkVVMScwJQYDVQQKEx5BQyBDYW1lcmZpcm1hIFNBIENJRiBBODI3NDMyODcxIzAhBgNVBAsTGmh0dHA6Ly93d3cuY2hhbWJlcnNpZ24ub3JnMSIwIAYDVQQDExlDaGFtYmVycyBvZiBDb21tZXJjZSBSb290MIIBIDANBgkqhkiG9w0BAQEFAAOCAQ0AMIIBCAKCAQEAtzZV5aVdGDDg2olUkfzIx1L4L1DZ77F1c2VHfRtbunXF/KGIJPov7coISjlUxFF6tdpg6jg8gbLL8bvZkSM/SAFwdakFKq0fcfPJVD0dBmpAPrMMhe5cG3nCYsS4No41XQEMIwRHNaqbYE6gZj3LJgqcQKH0XZi/caulAGgq7YN6D6IUtdQis4CwPAxaUWktWBiP7Zme8a7ileb2R6jWDA+wWFjbw2Y3npuRVDM30pQcakjJyfKl2qUMI/cjDpwyVV5xnIQFUZot/eZOKjRa3spAN2cMVCFVd9oKDMyXroDclDZK9D7ONhMeU+SsTjoF7Nuucpw4i9A5O4kKPnf+dQIBA6OCAUQwggFAMBIGA1UdEwEB/wQIMAYBAf8CAQwwPAYDVR0fBDUwMzAxoC+gLYYraHR0cDovL2NybC5jaGFtYmVyc2lnbi5vcmcvY2hhbWJlcnNyb290LmNybDAdBgNVHQ4EFgQU45T1sU3p26EpW1eLTXYGduHRooowDgYDVR0PAQH/BAQDAgEGMBEGCWCGSAGG+EIBAQQEAwIABzAnBgNVHREEIDAegRxjaGFtYmVyc3Jvb3RAY2hhbWJlcnNpZ24ub3JnMCcGA1UdEgQgMB6BHGNoYW1iZXJzcm9vdEBjaGFtYmVyc2lnbi5vcmcwWAYDVR0gBFEwTzBNBgsrBgEEAYGHLgoDATA+MDwGCCsGAQUFBwIBFjBodHRwOi8vY3BzLmNoYW1iZXJzaWduLm9yZy9jcHMvY2hhbWJlcnNyb290Lmh0bWwwDQYJKoZIhvcNAQEFBQADggEBAAxBl8IahsAifJ/7kPMa0QOx7xP5IV8EnNrJpY0nbJaHkb5BkAFyk+cefV/2icZdp0AJPaxJRUXcLo0waLIJuvvDL8y6C98/d3tGfToSJI6WjzwFCm/SlCgdbQzALogi1djPHRPH8EjX1wWnz8dHnjs8NMiAT9QUu/wNUPf6s+xCX6ndbcj0dc97wXImsQEcXCz9ek60AcUFV7nnPKoF2YjpB0ZBzu9Bga5Y34OirsrXdx/nADydb47kMgkdTXg0eDQ8lJsm7U9xxhl6vSAiSFr+S30Dt+dYvsYyTnQeaN2oaFuzPu5ifdmA6Ap1erfutGWaIZDgqtCYvDi1czyL+Nw="}}}}}},"responses":{"200":{"description":"TrustAnchor or root for chain was found in the registry"},"400":{"description":"Invalid request payload"},"404":{"description":"TrustAnchor was not be found in the registry"}},"tags":["TrustAnchor"]}},"/api/v2204/trustAnchor/chain":{"post":{"operationId":"TrustAnchorController_verifyTrustAnchorChainRaw","summary":"Verify root of a certificate chain to be a TrustAnchor in the registry","parameters":[],"requestBody":{"required":true,"description":"Verify root of a certificate chain to be a TrustAnchor in the registry","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrustAnchorChainRequestDto"},"examples":{"certChain":{"summary":"Example Certificate Chain","value":{"certs":"-----BEGIN CERTIFICATE-----MIIFKTCCBBGgAwIBAgISBD1oWylsswqAkAR9kUIZniXTMA0GCSqGSIb3DQEBCwUAMDIxCzAJBgNVBAYTAlVTMRYwFAYDVQQKEw1MZXQncyBFbmNyeXB0MQswCQYDVQQDEwJSMzAeFw0yMjA1MjAxMzU4MjlaFw0yMjA4MTgxMzU4MjhaMB0xGzAZBgNVBAMTEnJlZ2lzdHJ5LmdhaWEteC5ldTCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBANPRNchFiXS3nNCopDvI6bH2XHO9DatZg1AmGeEPhA5vrzbj0Y1Pr6V/WLQvABMtRK1xKlQHoTJMTTDEgKerzMG39tfOGE7wgz6oNh16tSU8YwvduC2+BeC880EH0eiqwIbtlCkY5ux3tVX57XQPwtV/aXWUfvqILqR7naQW0RBkBC3cvfgW001CYRilf60TJm+nlhGzgn5Y5QMf6xEotEpQjGpCF+EbdCptCpzPhozmh3ir67FzOT4fAsb6k61DlsuYrp33cugyv7snPjD6y9+4beDjsGZpGgy00SzmdaXvNgOL8Ip0r4AkHcvL45st04x+unzsKkmq0VkZJ6FHOjUCAwEAAaOCAkwwggJIMA4GA1UdDwEB/wQEAwIFoDAdBgNVHSUEFjAUBggrBgEFBQcDAQYIKwYBBQUHAwIwDAYDVR0TAQH/BAIwADAdBgNVHQ4EFgQU4Y5c2fUOLix6/BjdkBXMoQWpMtMwHwYDVR0jBBgwFoAUFC6zF7dYVsuuUAlA5h+vnYsUwsYwVQYIKwYBBQUHAQEESTBHMCEGCCsGAQUFBzABhhVodHRwOi8vcjMuby5sZW5jci5vcmcwIgYIKwYBBQUHMAKGFmh0dHA6Ly9yMy5pLmxlbmNyLm9yZy8wHQYDVR0RBBYwFIIScmVnaXN0cnkuZ2FpYS14LmV1MEwGA1UdIARFMEMwCAYGZ4EMAQIBMDcGCysGAQQBgt8TAQEBMCgwJgYIKwYBBQUHAgEWGmh0dHA6Ly9jcHMubGV0c2VuY3J5cHQub3JnMIIBAwYKKwYBBAHWeQIEAgSB9ASB8QDvAHUAQcjKsd8iRkoQxqE6CUKHXk4xixsD6+tLx2jwkGKWBvYAAAGA4foqgQAABAMARjBEAiEA7UgiGF3DOH96JdSND7tfC0o/kDKHPVlPSHerNKYGou8CHz8bV0VATgL9VNMVLTzbQWgmMAUqg3FnsILSfBkiPhQAdgBGpVXrdfqRIDC1oolp9PN9ESxBdL79SbiFq/L8cP5tRwAAAYDh+ip/AAAEAwBHMEUCIQDQc+rObD5KLDqBgmMMbxfPhF3Qtw6VMikDkbon0zzzfQIgB1VaR5oe9uyfNPpuaUrztF2QJJM+nwbFvTQtvvg1UqMwDQYJKoZIhvcNAQELBQADggEBACkWTJY5FtTsrvtUdFowAaPvZVnz+iuN2lYgTc+asgpfujiyoJWw/IQUVBpWTM15gr4qoY7QNjN20j8DGDS71wOoFp/SJ4wUEwiQdbCvO5swQrdXYkreWtakQgpEgXUOHLPHsFwjzqXMWoFjO+tvbAiUf7A2IyECa8p5jml/UpcBUkDp0qQiMTE2pBtp06lDj6I+Sf4aryqRoSyJ+/Cf+3kKLgy5o2FgVjjUByArfA9ydVo0XstUZhyf6JsarzTmXwWoCFYOt+uCGhwOzJ6ySIwBz15rYnUFVhWN5iyhBaD6BU4R3P0gnvBpPu+DCgasv4/orJtjQ/F/i5cRW35tqtM=-----END CERTIFICATE----------BEGIN CERTIFICATE-----MIIFFjCCAv6gAwIBAgIRAJErCErPDBinU/bWLiWnX1owDQYJKoZIhvcNAQELBQAwTzELMAkGA1UEBhMCVVMxKTAnBgNVBAoTIEludGVybmV0IFNlY3VyaXR5IFJlc2VhcmNoIEdyb3VwMRUwEwYDVQQDEwxJU1JHIFJvb3QgWDEwHhcNMjAwOTA0MDAwMDAwWhcNMjUwOTE1MTYwMDAwWjAyMQswCQYDVQQGEwJVUzEWMBQGA1UEChMNTGV0J3MgRW5jcnlwdDELMAkGA1UEAxMCUjMwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQC7AhUozPaglNMPEuyNVZLD+ILxmaZ6QoinXSaqtSu5xUyxr45r+XXIo9cPR5QUVTVXjJ6oojkZ9YI8QqlObvU7wy7bjcCwXPNZOOftz2nwWgsbvsCUJCWH+jdxsxPnHKzhm+/b5DtFUkWWqcFTzjTIUu61ru2P3mBw4qVUq7ZtDpelQDRrK9O8ZutmNHz6a4uPVymZ+DAXXbpyb/uBxa3Shlg9F8fnCbvxK/eG3MHacV3URuPMrSXBiLxgZ3Vms/EY96Jc5lP/Ooi2R6X/ExjqmAl3P51T+c8B5fWmcBcUr2Ok/5mzk53cU6cG/kiFHaFpriV1uxPMUgP17VGhi9sVAgMBAAGjggEIMIIBBDAOBgNVHQ8BAf8EBAMCAYYwHQYDVR0lBBYwFAYIKwYBBQUHAwIGCCsGAQUFBwMBMBIGA1UdEwEB/wQIMAYBAf8CAQAwHQYDVR0OBBYEFBQusxe3WFbLrlAJQOYfr52LFMLGMB8GA1UdIwQYMBaAFHm0WeZ7tuXkAXOACIjIGlj26ZtuMDIGCCsGAQUFBwEBBCYwJDAiBggrBgEFBQcwAoYWaHR0cDovL3gxLmkubGVuY3Iub3JnLzAnBgNVHR8EIDAeMBygGqAYhhZodHRwOi8veDEuYy5sZW5jci5vcmcvMCIGA1UdIAQbMBkwCAYGZ4EMAQIBMA0GCysGAQQBgt8TAQEBMA0GCSqGSIb3DQEBCwUAA4ICAQCFyk5HPqP3hUSFvNVneLKYY611TR6WPTNlclQtgaDqw+34IL9fzLdwALduO/ZelN7kIJ+m74uyA+eitRY8kc607TkC53wlikfmZW4/RvTZ8M6UK+5UzhK8jCdLuMGYL6KvzXGRSgi3yLgjewQtCPkIVz6D2QQzCkcheAmCJ8MqyJu5zlzyZMjAvnnAT45tRAxekrsu94sQ4egdRCnbWSDtY7kh+BImlJNXoB1lBMEKIq4QDUOXoRgffuDghje1WrG9ML+Hbisq/yFOGwXD9RiX8F6sw6W4avAuvDszue5L3sz85K+EC4Y/wFVDNvZo4TYXao6Z0f+lQKc0t8DQYzk1OXVu8rp2yJMC6alLbBfODALZvYH7n7do1AZls4I9d1P4jnkDrQoxB3UqQ9hVl3LEKQ73xF1OyK5GhDDX8oVfGKF5u+decIsH4YaTw7mP3GFxJSqv3+0lUFJoi5Lc5da149p90IdshCExroL1+7mryIkXPeFM5TgO9r0rvZaBFOvV2z0gp35Z0+L4WPlbuEjN/lxPFin+HlUjr8gRsI3qfJOQFy/9rKIJR0Y/8Omwt/8oTWgy1mdeHmmjk7j1nYsvC9JSQ6ZvMldlTTKB3zhThV1+XWYp6rjd5JW1zbVWEkLNxE7GJThEUG3szgBVGP7pSWTUTsqXnLRbwHOoq7hHwg==-----END CERTIFICATE----------BEGIN CERTIFICATE-----MIIFazCCA1OgAwIBAgIRAIIQz7DSQONZRGPgu2OCiwAwDQYJKoZIhvcNAQELBQAwTzELMAkGA1UEBhMCVVMxKTAnBgNVBAoTIEludGVybmV0IFNlY3VyaXR5IFJlc2VhcmNoIEdyb3VwMRUwEwYDVQQDEwxJU1JHIFJvb3QgWDEwHhcNMTUwNjA0MTEwNDM4WhcNMzUwNjA0MTEwNDM4WjBPMQswCQYDVQQGEwJVUzEpMCcGA1UEChMgSW50ZXJuZXQgU2VjdXJpdHkgUmVzZWFyY2ggR3JvdXAxFTATBgNVBAMTDElTUkcgUm9vdCBYMTCCAiIwDQYJKoZIhvcNAQEBBQADggIPADCCAgoCggIBAK3oJHP0FDfzm54rVygch77ct984kIxuPOZXoHj3dcKi/vVqbvYATyjb3miGbESTtrFj/RQSa78f0uoxmyF+0TM8ukj13Xnfs7j/EvEhmkvBioZxaUpmZmyPfjxwv60pIgbz5MDmgK7iS4+3mX6UA5/TR5d8mUgjU+g4rk8Kb4Mu0UlXjIB0ttov0DiNewNwIRt18jA8+o+u3dpjq+sWT8KOEUt+zwvo/7V3LvSye0rgTBIlDHCNAymg4VMk7BPZ7hm/ELNKjD+Jo2FR3qyHB5T0Y3HsLuJvW5iB4YlcNHlsdu87kGJ55tukmi8mxdAQ4Q7e2RCOFvu396j3x+UCB5iPNgiV5+I3lg02dZ77DnKxHZu8A/lJBdiB3QW0KtZB6awBdpUKD9jf1b0SHzUvKBds0pjBqAlkd25HN7rOrFleaJ1/ctaJxQZBKT5ZPt0m9STJEadao0xAH0ahmbWnOlFuhjuefXKnEgV4We0+UXgVCwOPjdAvBbI+e0ocS3MFEvzG6uBQE3xDk3SzynTnjh8BCNAw1FtxNrQHusEwMFxIt4I7mKZ9YIqioymCzLq9gwQbooMDQaHWBfEbwrbwqHyGO0aoSCqI3Haadr8faqU9GY/rOPNk3sgrDQoo//fb4hVC1CLQJ13hef4Y53CIrU7m2Ys6xt0nUW7/vGT1M0NPAgMBAAGjQjBAMA4GA1UdDwEB/wQEAwIBBjAPBgNVHRMBAf8EBTADAQH/MB0GA1UdDgQWBBR5tFnme7bl5AFzgAiIyBpY9umbbjANBgkqhkiG9w0BAQsFAAOCAgEAVR9YqbyyqFDQDLHYGmkgJykIrGF1XIpu+ILlaS/V9lZLubhzEFnTIZd+50xx+7LSYK05qAvqFyFWhfFQDlnrzuBZ6brJFe+GnY+EgPbk6ZGQ3BebYhtF8GaV0nxvwuo77x/Py9auJ/GpsMiu/X1+mvoiBOv/2X/qkSsisRcOj/KKNFtY2PwByVS5uCbMiogziUwthDyC3+6WVwW6LLv3xLfHTjuCvjHIInNzktHCgKQ5ORAzI4JMPJ+GslWYHb4phowim57iaztXOoJwTdwJx4nLCgdNbOhdjsnvzqvHu7UrTkXWStAmzOVyyghqpZXjFaH3pO3JLF+l+/+sKAIuvtd7u+Nxe5AW0wdeRlN8NwdCjNPElpzVmbUq4JUagEiuTDkHzsxHpFKVK7q4+63SM1N95R1NbdWhscdCb+ZAJzVcoyi3B43njTOQ5yOf+1CceWxG1bQVs5ZufpsMljq4Ui0/1lvh+wjChP4kqKOJ2qxq4RgqsahDYVvTH9w7jXbyLeiNdd8XM2w9U/t7y0Ff/9yi0GE44Za4rF2LN9d11TPAmRGunUHBcnWEvgJBQl9nJEiU0Zsnvgc/ubhPgXRR4Xq37Z0j4r7g1SgEEzwxA57demyPxgcYxn/eR44/KJ4EBs+lVDR3veyJm+kXQ99b21/+jh5Xos1AnX5iItreGCc=-----END CERTIFICATE-----"}}}}}},"responses":{"200":{"description":"TrustAnchor or root for chain was found in the registry"},"400":{"description":"Invalid request payload"},"409":{"description":"Root for the certificate chain could not be verified as a TrustAnchor in the registry"}},"tags":["TrustAnchor"]}},"/api/v2204/trustAnchor/chain/file":{"post":{"operationId":"TrustAnchorController_verifyTrustAnchorChain","summary":"Verify root of a certificate chain, provided as a file at uri, to be a TrustAnchor in the registry","parameters":[],"requestBody":{"required":true,"description":"Verify root of a certificate chain, provided as a file at uri, to be a TrustAnchor in the registry","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrustAnchorChainUriRequestDto"},"examples":{"certChain":{"summary":"Example Certificate Chain","value":{"uri":"https://compliance.gaia-x.eu/.well-known/x509CertificateChain.pem"}}}}}},"responses":{"200":{"description":"TrustAnchor or root for chain was found in the registry"},"400":{"description":"Invalid request payload"},"409":{"description":"Root for the certificate chain could not be verified as a TrustAnchor in the registry"}},"tags":["TrustAnchor"]}},"/api/v2204/termsAndConditions":{"get":{"operationId":"TermsAndConditionsController_getTermsAndConditions","summary":"Get terms and conditions for specified version or a list of available versions when no version is specified.","parameters":[{"name":"version","required":false,"in":"query","schema":{"enum":["22.04"],"type":"string"}}],"responses":{"200":{"description":"The version, terms and conditions text and the sha256 of the text."},"400":{"description":"Invalid request query"}},"tags":["TermsAndConditions"]}},"/api/v2204/shape":{"get":{"operationId":"ShapeController_getContext","summary":"Get a JSONLD context for all available shapes.","parameters":[],"responses":{"200":{"description":"The JSONLD context for all available shapes in the registry."}},"tags":["Shape"]}},"/api/v2204/shape/files":{"get":{"operationId":"ShapeController_getShape","summary":"Get specified SHACL file as ttl or jsonld or a list of all available shapes if no file and type is specified.","parameters":[{"name":"file","required":false,"in":"query","schema":{"enum":["participant","resource","service-offering"],"type":"string"}},{"name":"type","required":false,"in":"query","schema":{"enum":["ttl","jsonld"],"type":"string"}}],"responses":{"200":{"description":"SHACL file as ttl or jsonld or a list of available files when no file is specified."},"400":{"description":"Invalid request query"}},"tags":["Shape"]}},"/api/v2206/shape":{"get":{"operationId":"ShapeController2206_getContext","summary":"Get a JSONLD context for all available shapes.","parameters":[],"responses":{"200":{"description":"The JSONLD context for all available shapes in the registry."}},"tags":["Shape"]}},"/api/v2206/shape/files":{"get":{"operationId":"ShapeController2206_getShape","summary":"Get specified SHACL file as ttl or jsonld or a list of all available shapes if no file and type is specified.","parameters":[{"name":"file","required":false,"in":"query","schema":{"enum":["participant","resource","service-offering"],"type":"string"}},{"name":"type","required":false,"in":"query","schema":{"enum":["ttl","jsonld"],"type":"string"}}],"responses":{"200":{"description":"SHACL file as ttl or jsonld or a list of available files when no file is specified."},"400":{"description":"Invalid request query"}},"tags":["Shape"]}}},"info":{"title":"gx-registry","description":"Gaia-X Lab Registry","version":"1.1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"TrustAnchorRequestDto":{"type":"object","properties":{"certificate":{"type":"string","description":"The raw certificate in pem format"}},"required":["certificate"]},"TrustAnchorChainRequestDto":{"type":"object","properties":{"certs":{"type":"string","description":"The raw certificate chain in pem format"}},"required":["certs"]},"TrustAnchorChainUriRequestDto":{"type":"object","properties":{"uri":{"type":"string","description":"The uri to the file containing the certificate chain. PEM or PKCS7 format are expected."}},"required":["uri"]}}}}